Split the remove measurements pass into its own source file.
Apply pass in parallel.

Remove dead code.
