+++ lib/Optimizer/CodeGen/CMakeLists.txt
@@ -14,6 +14,7 @@ add_cudaq_library(OptCodeGen
   LowerToQIRProfile.cpp
   LowerToQIR.cpp
   Passes.cpp
+  RemoveMeasurements.cpp
   TranslateToIQMJson.cpp
   TranslateToOpenQASM.cpp
