+++ lib/Optimizer/CodeGen/LowerToQIRProfile.cpp
@@ -654,51 +654,3 @@ void cudaq::opt::addQIRProfilePipeline(OpPassManager &pm,
   pm.addPass(createQIRToQIRProfilePass(convertTo));
   pm.addNestedPass<LLVM::LLVMFuncOp>(verifyQIRProfilePass(convertTo));
 }
-
-namespace cudaq {
-
-struct EraseMeasurements : public OpRewritePattern<LLVM::CallOp> {
-  using OpRewritePattern<LLVM::CallOp>::OpRewritePattern;
-
-  void initialize() { setDebugName("EraseMeasurements"); }
-
-  LogicalResult matchAndRewrite(LLVM::CallOp call,
-                                PatternRewriter &rewriter) const override {
-    if (auto callee = call.getCallee()) {
-      if (callee->equals(cudaq::opt::QIRMeasureBody) ||
-          callee->equals(cudaq::opt::QIRRecordOutput)) {
-        rewriter.eraseOp(call);
-        return success();
-      }
-    }
-    return failure();
-  }
-};
-
-/// Remove Measurements
-///
-/// This pass removes measurements and the corresponding output recording calls.
-/// This is needed for backends that don't support selective measurement calls.
-/// For example: https://github.com/NVIDIA/cuda-quantum/issues/512
-struct RemoveMeasurementsPass
-    : public cudaq::opt::RemoveMeasurementsBase<RemoveMeasurementsPass> {
-  explicit RemoveMeasurementsPass() = default;
-
-  void runOnOperation() override {
-    auto *op = getOperation();
-    auto *context = &getContext();
-    RewritePatternSet patterns(context);
-    patterns.insert<EraseMeasurements>(context);
-    if (failed(applyPatternsAndFoldGreedily(op, std::move(patterns)))) {
-      signalPassFailure();
-    }
-  }
-
-private:
-  FrozenRewritePatternSet patterns;
-};
-} // namespace cudaq
-
-std::unique_ptr<Pass> cudaq::opt::createRemoveMeasurementsPass() {
-  return std::make_unique<RemoveMeasurementsPass>();
-}
