+++ runtime/cudaq/platform/default/rest/helpers/ionq/ionq.config
@@ -23,7 +23,7 @@ CODEGEN_EMISSION=qir-base
 
 # Additional passes to run after lowering to QIR
 # This is required due to https://github.com/NVIDIA/cuda-quantum/issues/512
-POST_CODEGEN_PASSES="remove-measurements"
+POST_CODEGEN_PASSES="func.func(remove-measurements),llvm.llvmfunc(remove-measurements)"
 
 # Library mode is only for simulators, 
 # and it is the default, physical backends must
